Move waste-balances domain and application into module
