+++ src/domain/waste-records/model.js
@@ -53,6 +53,6 @@ export const WASTE_RECORD_CHANGE = Object.freeze({
  * @property {WasteRecordType} type
  * @property {Object} data - Reporting fields only
  * @property {WasteRecordVersion[]} versions - Version history. createdAt/updatedAt derived from first/last version
- * @property {import('#domain/waste-balances/model.js').UserSummary} [updatedBy] - User who last updated the record
+ * @property {import('#waste-balances/domain/model.js').UserSummary} [updatedBy] - User who last updated the record
  * @property {boolean} [excludedFromWasteBalance] - Set by markExcludedRecords when the row fails schema validation
  */
