+++ src/waste-balances/repository/contract/deductAvailableBalanceForPrnCreation.contract.js
@@ -1,6 +1,6 @@
 import { describe, beforeEach, expect } from 'vitest'
 import { buildWasteBalance } from './test-data.js'
-import { WASTE_BALANCE_TRANSACTION_ENTITY_TYPE } from '#domain/waste-balances/model.js'
+import { WASTE_BALANCE_TRANSACTION_ENTITY_TYPE } from '../../domain/model.js'
 
 export const testDeductAvailableBalanceForPrnCreationBehaviour = (it) => {
   describe('deductAvailableBalanceForPrnCreation', () => {
